% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.ref.R
\docType{data}
\name{coord.ref}
\alias{coord.ref}
\title{Object for illustrating documentation, cf vignette}
\format{a data.frame}
\source{
simulated
}
\usage{
coord.ref
}
\description{
Object for illustrating documentation, cf vignette
}
\details{
coordinates
}
\keyword{datasets}
