% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.ref.R
\docType{data}
\name{geno.ref}
\alias{geno.ref}
\title{Object for illustrating documentation, cf vignette}
\format{a matrix}
\source{
simulated
}
\usage{
geno.ref
}
\description{
Object for illustrating documentation, cf vignette
}
\details{
allele counts for 50 populations at 500 loci
}
\keyword{datasets}
