% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.unknown.R
\docType{data}
\name{geno.unknown}
\alias{geno.unknown}
\title{Object for illustrating documentation, cf vignette}
\format{a matrix}
\source{
simulated
}
\usage{
geno.unknown
}
\description{
Object for illustrating documentation, cf vignette
}
\details{
genotypes (allele counts) for 20 diploid individuals at 500 loci
}
\keyword{datasets}
