% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.pop.ref.R
\docType{data}
\name{size.pop.ref}
\alias{size.pop.ref}
\title{Object for illustrating documentation, cf vignette}
\format{a data.frame}
\source{
simulated
}
\usage{
size.pop.ref
}
\description{
Object for illustrating documentation, cf vignette
}
\details{
coordinates of individuals  whose genotypes are contained in matrix `geno.unknwon'
(assumed in the vignette examples to be of unknown geographic origin)
}
\keyword{datasets}
