% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPASIBA.sim.R
\name{SPASIBA.sim}
\alias{SPASIBA.sim}
\title{Simulate data from the prior/likelihood model assumed in SPASIBA}
\usage{
SPASIBA.sim(
  ploidy = 2,
  n.loc = 10,
  n.quant = 0,
  sphere = FALSE,
  size.pop.ref,
  coloc.ref = FALSE,
  n.unknown,
  nx.grid,
  ny.grid,
  win,
  kappa.geno = NULL,
  nu.geno = 1,
  sigma2.geno = NULL,
  kappa.quant = NULL,
  nu.quant = 1,
  sigma2.gf.quant = NULL,
  marginal.quant = "lnorm",
  tau.noise.quant = NULL,
  store.gf = FALSE,
  obsolete = FALSE
)
}
\arguments{
\item{ploidy}{1 or 2}

\item{n.loc}{integer nuber of loci}

\item{n.quant}{number of quantitative covariates}

\item{size.pop.ref}{Matrix with one row per population and one column 
per locus giving haploid population sample size}

\item{coloc.ref}{TRUE / FALSE indicating whether colocated samples are allowed}

\item{n.unknown}{Number of individuls with unknown spatial coordinates}

\item{nx.grid}{Number of pixels horizontally}

\item{ny.grid}{Number of pixels vertically}

\item{win}{c(x.min,x.max,y.min,y.max)}

\item{kappa.geno}{spatial scale parameter, hidden Gaussian RF, genotypes}

\item{nu.geno}{spatial smoothing parameter, hidden Gaussian RF, genotypes}

\item{sigma2.geno}{marginal variance, hidden Gaussian RF, genotypes}

\item{kappa.quant}{spatial scale parameter, quant  variables}

\item{nu.quant}{spatial smoothing parameter, quant  variables}

\item{sigma2.gf.quant}{marginal variance, quant  variables}

\item{marginal.quant}{distribution of quantitative variables ('norm' or 'lnorm')}

\item{tau.noise.quant}{precision of distribution of quantitative variables}

\item{store.gf}{TRUE / FALSE indicating whether Gaussian random field 
values used to simulated allele frequencies should be stored}

\item{obsolete}{TRUE / FALSE indicating whether random field 
simulations should be done with the obsolete function GaussRF or the more recent function RFsimulate}

\item{flat}{domai}
}
\value{
A list with components 
\itemize{
\item sim.method: The simulation method. The function currently
  implements the "geostat" model
 \item sphere: A logical indicating whether coordinates are handled as
  (Lat,Lon) on the sphere or Euclidean coordinates in a 2D flat domain
 \item geno.ref: A matrix containing the genetic data (individual or 
 \item population allele counts)  of the reference 
samples. This matrix has one line per population (or individual) and one column per locus
 \item coord.ref: A matrix containing the geographic coordinates of 
 \item reference populations. This matrix has one line per population (or individual) and two columns
 \item size.pop.ref: A matrix containing the haploid population size of reference populations. 
This matrix has one line per population (or individual) and one column per locus
 \item geno.unknown: Indivudual genotypes of individuals of unknown geographic origin
 \item true.coord.unknown: True geographic coordinates of individuals assumed here to be of 
unknown geographic origin
 \item ploidy: 1 or 2 
 \item kappa.geno: Scale parameter used in the simulation of the underlying Matern field 
 \item nu.geno: Smoothing parameter used in the simulation of the underlying Matern field
 \item sigma2.geno: Variance parameter used in the simulation of the underlying Matern field
 \item coord.grid=coord: Coordinates of the regular grid at which simulation of 
Gaussian fields and allele frequencies is performed
}
}
\description{
Simulate data from the prior/likelihood model assumed in SPASIBA
}
\examples{
dat <- SPASIBA.sim(ploidy=2,
                   n.loc=3, # nb of loci
                   sphere=FALSE,
                   size.pop.ref=rep(2,30), # haploid pop size at each site
                   n.unknown=20, # nb indiv with unknown origin
                   nx.grid=100, # nb hor. pixels on grid 
                   ny.grid=100, # nb vert.  pixels on grid 
                   win=c(0,10,40,50),
                   kappa.geno=1/10,# spatial scale parameter
                   nu.geno=1,# spatial smoothing parameter
                   sigma2.geno=1# marginal variance of hidden Gaussian RF
                   )
                   summary(dat)
}
\author{
Gilles Guillot
}
