% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true.coord.unknown.R
\docType{data}
\name{true.coord.unknown}
\alias{true.coord.unknown}
\title{Object for illustrating documentation, cf vignette}
\format{a data.frame}
\source{
simulated
}
\usage{
true.coord.unknown
}
\description{
Object for illustrating documentation, cf vignette
}
\details{
haploid population size for 50 populations at 500 loci
}
\keyword{datasets}
